requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Redis (for caching & rate limiting)
redis==5.0.1
Flask-Caching==2.1.0

# Computer Vision
ultralytics==8.1.0
opencv-python-headless==4.8.1.78
Pillow==10.1.0

# API & Data Processing
requests==2.31.0
python-dotenv==1.0.0

# Security & Validation
Werkzeug==3.0.1
email-validator==2.1.0
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Production Server
gunicorn==21.2.0